Add entry point for package when installed. Should help with #6
